Unify BASIC runtime symbol resolution
