+++ basic/src/basicc.c
@@ -71,10 +71,7 @@ static void basic_mir_bcmp (MIR_context_t ctx, MIR_item_t func, MIR_insn_code_t 
 #include <stdint.h>
 #include <stdarg.h>
 #include <unistd.h>
-#include <dlfcn.h>
 #include <math.h>
-#include "mir-hash.h"
-
 #ifndef BASIC_SRC_DIR
 #define BASIC_SRC_DIR "."
 #endif
@@ -142,178 +139,9 @@ static int safe_snprintf (char *buf, size_t size, const char *fmt, ...) {
   return res;
 }
 
-static BasicRuntimeSymbol runtime_symbols[] = {
-  {"basic_print", basic_print},
-  {"basic_print_str", basic_print_str},
-  {"basic_input", basic_input},
-  {"basic_input_str", basic_input_str},
-  {"basic_get", basic_get},
-  {"basic_inkey", basic_inkey},
-  {"basic_put", basic_put},
-  {"basic_strcmp", basic_strcmp},
-  {"basic_open", basic_open},
-  {"basic_close", basic_close},
-  {"basic_print_hash", basic_print_hash},
-  {"basic_print_hash_str", basic_print_hash_str},
-  {"basic_input_hash", basic_input_hash},
-  {"basic_input_hash_str", basic_input_hash_str},
-  {"basic_get_hash", basic_get_hash},
-  {"basic_put_hash", basic_put_hash},
-  {"basic_eval", basic_eval},
-  {"basic_eof", basic_eof},
-  {"basic_read", basic_read},
-  {"basic_read_str", basic_read_str},
-  {"basic_restore", basic_restore},
-  {"basic_clear_array", basic_clear_array},
-  {"basic_dim_alloc", basic_dim_alloc},
-  {"basic_home", basic_home},
-  {"basic_vtab", basic_vtab},
-  {"basic_randomize", basic_randomize},
-  {"basic_rnd", basic_rnd},
-  {"basic_abs", basic_abs},
-  {"basic_sgn", basic_sgn},
-  {"basic_iabs", basic_iabs},
-  {"basic_isgn", basic_isgn},
-  {"basic_sqr", basic_sqr},
-  {"basic_sin", basic_sin},
-  {"basic_cos", basic_cos},
-  {"basic_tan", basic_tan},
-  {"basic_atn", basic_atn},
-  {"basic_sinh", basic_sinh},
-  {"basic_cosh", basic_cosh},
-  {"basic_tanh", basic_tanh},
-  {"basic_asinh", basic_asinh},
-  {"basic_acosh", basic_acosh},
-  {"basic_atanh", basic_atanh},
-  {"basic_asin", basic_asin},
-  {"basic_acos", basic_acos},
-  {"basic_log", basic_log},
-  {"basic_log2", basic_log2},
-  {"basic_log10", basic_log10},
-  {"basic_exp", basic_exp},
-  {"basic_pow", basic_pow},
-  {"basic_pi", basic_pi},
-  {"basic_instr", basic_instr},
-  {"basic_screen", basic_screen},
-  {"basic_cls", basic_cls},
-  {"basic_color", basic_color},
-  {"basic_key_off", basic_key_off},
-  {"basic_locate", basic_locate},
-  {"basic_tab", basic_tab},
-  {"basic_htab", basic_htab},
-  {"basic_pos", basic_pos},
-  {"basic_text", basic_text},
-  {"basic_inverse", basic_inverse},
-  {"basic_normal", basic_normal},
-  {"basic_hgr2", basic_hgr2},
-  {"basic_hcolor", basic_hcolor},
-  {"basic_hplot", basic_hplot},
-  {"basic_hplot_to", basic_hplot_to},
-  {"basic_hplot_to_current", basic_hplot_to_current},
-  {"basic_move", basic_move},
-  {"basic_draw", basic_draw},
-  {"basic_draw_line", basic_draw_line},
-  {"basic_circle", basic_circle},
-  {"basic_rect", basic_rect},
-  {"basic_fill", basic_fill},
-  {"basic_mode", basic_mode},
-  {"basic_profile_line", basic_profile_line},
-  {"basic_profile_func_enter", basic_profile_func_enter},
-  {"basic_profile_func_exit", basic_profile_func_exit},
-  {"basic_chr", basic_chr},
-  {"basic_unichar", basic_unichar},
-  {"basic_string", basic_string},
-  {"basic_concat", basic_concat},
-  {"basic_left", basic_left},
-  {"basic_right", basic_right},
-  {"basic_mid", basic_mid},
-  {"basic_mirror", basic_mirror},
-  {"basic_upper", basic_upper},
-  {"basic_lower", basic_lower},
-  {"basic_len", basic_len},
-  {"basic_val", basic_val},
-  {"basic_str", basic_str},
-  {"basic_asc", basic_asc},
-  {"basic_int", basic_int},
-  {"basic_timer", basic_timer},
-  {"basic_time_str", basic_time_str},
-  {"basic_date_str", basic_date_str},
-  {"basic_input_chr", basic_input_chr},
-  {"basic_peek", basic_peek},
-  {"basic_poke", basic_poke},
-  {"basic_stop", basic_stop},
-  {"basic_return_error", basic_return_error},
+static BasicRuntimeSymbol local_runtime_symbols[] = {
   {"basic_chain", basic_chain},
-  {"basic_set_error_handler", basic_set_error_handler},
-  {"basic_get_error_handler", basic_get_error_handler},
-  {"basic_set_line", basic_set_line},
-  {"basic_get_line", basic_get_line},
-  {"basic_enable_line_tracking", basic_enable_line_tracking},
-  {"basic_delay", basic_delay},
-  {"basic_beep", basic_beep},
-  {"basic_sound", basic_sound},
-  {"basic_sound_off", basic_sound_off},
-  {"basic_system", basic_system},
-  {"basic_system_out", basic_system_out},
-  {"basic_strdup", basic_strdup},
-  {"basic_free", basic_free},
-  {"basic_calloc", basic_calloc},
-  {"basic_pool_reset", basic_pool_reset},
-  {"memset", memset},
-  {"basic_mir_ctx", basic_mir_ctx},
-  {"basic_mir_mod", basic_mir_mod},
-  {"basic_mir_func", basic_mir_func},
-  {"basic_mir_reg", basic_mir_reg},
-  {"basic_mir_label", basic_mir_label},
-  {"basic_mir_emit", basic_mir_emit},
-  {"basic_mir_emitlbl", basic_mir_emitlbl},
-  {"basic_mir_ret", basic_mir_ret},
-  {"basic_mir_finish", basic_mir_finish},
-  {"basic_mir_run", basic_mir_run},
-  {"basic_mir_dump", basic_mir_dump},
-  {"basic_fact", basic_fact},
 };
-
-#define RUNTIME_SYMBOLS_CNT (sizeof (runtime_symbols) / sizeof (runtime_symbols[0]))
-
-#define RUNTIME_HASH_SIZE 256
-static BasicRuntimeSymbol *runtime_symbol_hash[RUNTIME_HASH_SIZE];
-static int runtime_symbol_hash_inited = 0;
-
-static size_t runtime_symbol_hash_func (const char *s) {
-  return (size_t) mir_hash (s, strlen (s), 0);
-}
-
-static void runtime_symbol_hash_init (void) {
-  for (size_t i = 0; i < RUNTIME_SYMBOLS_CNT; ++i) {
-    size_t h = runtime_symbol_hash_func (runtime_symbols[i].name) & (RUNTIME_HASH_SIZE - 1);
-    while (runtime_symbol_hash[h] != NULL) h = (h + 1) & (RUNTIME_HASH_SIZE - 1);
-    runtime_symbol_hash[h] = &runtime_symbols[i];
-  }
-  runtime_symbol_hash_inited = 1;
-}
-
-static void *runtime_symbol_lookup (const char *name) {
-  if (!runtime_symbol_hash_inited) runtime_symbol_hash_init ();
-  size_t h = runtime_symbol_hash_func (name) & (RUNTIME_HASH_SIZE - 1);
-  for (;;) {
-    BasicRuntimeSymbol *sym = runtime_symbol_hash[h];
-    if (sym == NULL) return NULL;
-    if (strcmp (name, sym->name) == 0) return sym->fn;
-    h = (h + 1) & (RUNTIME_HASH_SIZE - 1);
-  }
-}
-
-static void *resolve (const char *name) {
-  void *fn = runtime_symbol_lookup (name);
-  if (fn != NULL) return fn;
-  return dlsym (NULL, name);
-}
-
-size_t basic_runtime_symbols (BasicRuntimeSymbol **syms) {
-  if (syms != NULL) *syms = runtime_symbols;
-  return RUNTIME_SYMBOLS_CNT;
-}
 
 /* Runtime call prototypes for expressions */
 static MIR_item_t rnd_proto, rnd_import, chr_proto, chr_import, unichar_proto, unichar_import,
@@ -6898,7 +6726,7 @@ static void gen_program (LineVec *prog, int jit, int asm_p, int obj_p, int bin_p
   if (code_p) {
     MIR_load_module (ctx, module);
     MIR_gen_init (ctx);
-    MIR_link (ctx, MIR_set_gen_interface, resolve);
+    MIR_link (ctx, MIR_set_gen_interface, basic_runtime_resolve);
     for (size_t i = 0; i < func_defs.len; i++) {
       FuncDef *fd = &func_defs.data[i];
       if (fd->is_extern) continue;
@@ -6923,7 +6751,7 @@ static void gen_program (LineVec *prog, int jit, int asm_p, int obj_p, int bin_p
   MIR_load_module (ctx, module);
   if (jit) {
     MIR_gen_init (ctx);
-    MIR_link (ctx, MIR_set_gen_interface, resolve);
+    MIR_link (ctx, MIR_set_gen_interface, basic_runtime_resolve);
     for (size_t i = 0; i < func_defs.len; i++) {
       FuncDef *fd = &func_defs.data[i];
       if (fd->is_extern) continue;
@@ -6934,7 +6762,7 @@ static void gen_program (LineVec *prog, int jit, int asm_p, int obj_p, int bin_p
     m ();
     MIR_gen_finish (ctx);
   } else {
-    MIR_link (ctx, MIR_set_interp_interface, resolve);
+    MIR_link (ctx, MIR_set_interp_interface, basic_runtime_resolve);
     typedef int (*main_t) (void);
     main_t m = func->addr;
     m ();
@@ -7294,6 +7122,8 @@ int main (int argc, char **argv) {
 #endif
   arena_init (&ast_arena);
   basic_pool_reset ();
+  basic_runtime_add_symbols (local_runtime_symbols,
+                             sizeof (local_runtime_symbols) / sizeof (local_runtime_symbols[0]));
   load_repl_abbrevs (argv[0]);
   int jit = 0, asm_p = 0, obj_p = 0, bin_p = 0, reduce_libs = 0;
   const char *fname = NULL, *out_name = NULL;
